Refactored BioMCP to have a separate MCP server
