+++ bioagents/models/__init__.py
@@ -1,5 +1,21 @@
 from .llms import LLM
 from .source import Source
 from .file_info import FileInfo
+from .pubmed import (
+    PubMedArticle,
+    PubMedArticleDetails,
+    PubMedParser,
+    parse_articles,
+    parse_article_details,
+)
 
-__all__ = ["LLM", "Source", "FileInfo"]
+__all__ = [
+    "LLM",
+    "Source",
+    "FileInfo",
+    "PubMedArticle",
+    "PubMedArticleDetails",
+    "PubMedParser",
+    "parse_articles",
+    "parse_article_details",
+]
