+++ tests/conftest.py
@@ -5,8 +5,15 @@ Pytest configuration and fixtures for bioagents tests.
 import pytest
 import asyncio
 import os
+import sys
+from pathlib import Path
 from unittest.mock import AsyncMock, MagicMock
 from typing import AsyncGenerator
+
+# Ensure project root is on sys.path for imports like `bioagents.*`
+PROJECT_ROOT = Path(__file__).resolve().parents[1]
+if str(PROJECT_ROOT) not in sys.path:
+    sys.path.insert(0, str(PROJECT_ROOT))
 
 # Load environment variables from .env for testing
 try:
